refactor: extract shared crate coronarium-core for Linux↔Windows parity
Split the platform-neutral bits of the Linux coronarium binary into a
new crate `coronarium-core`:
- events::Event (schema)
- stats::Stats + ingestion with per-kind sampling caps
- policy (YAML loader, Mode, DefaultDecision, FilePolicy, ...)
- matcher (FileMatcher, ExecMatcher) — with the existing 5 unit tests
- html::render (self-contained dashboard)
- report::write (JSON log + GITHUB_STEP_SUMMARY + HTML)

The Linux bin now re-exports from core (events.rs keeps just the eBPF
ringbuf decode). coronarium-win uses the same core crate, so both
binaries emit byte-identical JSON logs and HTML reports.

Windows side (crates/coronarium-win) is now a proper audit tool:

- clap-parsed CLI matching Linux: --policy / --mode / --log / --summary
  / --html, trailing-var-arg child command.
- ETW callbacks extract real fields via ferrisetw's Parser and emit the
  same Event::{Exec,Connect,Open} as Linux.
  * Process provider event id 1 → Exec (ImageName, CommandLine, PID)
  * Network provider ids 12/14 → Connect (daddr, dport, PID)
  * File provider ids 12/30 → Open (FileName, PID) with \\ → /
    normalisation so policies written for Linux optionally match.
- Userspace ExecMatcher / FileMatcher applied identically, tagging
  `denied: true` and triggering the same `::error title=coronarium::`
  annotation + exit-1 in block mode.

Windows smoke workflow rewritten to exercise:
- audit-mode with --html + --log, asserting exec samples fire.
- block-mode with `process.deny_exec: [whoami]`, asserting a denied
  exec sample is present AND coronarium exits non-zero.
