+++ crates/coronarium-core/src/html.rs
@@ -1,0 +1,499 @@
+//! Self-contained HTML audit report.
+//!
+//! Single file, no external CSS/JS — we want the artifact to work straight
+//! from `file://` with no network, and remain trivially diff-able. Dark mode
+//! is auto-switched via `prefers-color-scheme`.
+
+use std::fmt::Write;
+
+use crate::{events::Event, policy::Policy, stats::Stats};
+
+/// Render the full report. The result is a complete `<!DOCTYPE html>`
+/// document.
+pub fn render(policy: &Policy, stats: &Stats, meta: ReportMeta<'_>) -> String {
+    let mut html = String::with_capacity(16 * 1024);
+
+    let by_kind = count_by_kind(&stats.samples);
+    let badge_class = if stats.denied > 0 {
+        "badge-danger"
+    } else if stats.lost > 0 {
+        "badge-warn"
+    } else {
+        "badge-ok"
+    };
+    let status_label = if stats.denied > 0 {
+        format!("{} denied", stats.denied)
+    } else if stats.lost > 0 {
+        "events dropped".to_string()
+    } else {
+        "clean".to_string()
+    };
+
+    let _ = write!(
+        html,
+        r#"<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>coronarium report — {title}</title>
+<style>{css}</style>
+</head>
+<body>
+<header class="top">
+  <div class="brand">
+    <span class="logo">🛡️</span>
+    <span class="title">coronarium</span>
+  </div>
+  <div class="badges">
+    <span class="badge mode-{mode_lc}">mode: {mode}</span>
+    <span class="badge {badge_class}">{status_label}</span>
+    <span class="badge">{observed} events</span>
+  </div>
+</header>
+
+<main>
+  <section class="stats">
+    <div class="card"><div class="card-label">observed</div><div class="card-value">{observed}</div></div>
+    <div class="card"><div class="card-label">denied</div><div class="card-value {denied_class}">{denied}</div></div>
+    <div class="card"><div class="card-label">lost</div><div class="card-value {lost_class}">{lost}</div></div>
+    <div class="card"><div class="card-label">kinds</div>
+      <div class="card-kinds">
+        <span class="chip chip-exec">exec · {n_exec}</span>
+        <span class="chip chip-open">open · {n_open}</span>
+        <span class="chip chip-connect">connect · {n_connect}</span>
+      </div>
+    </div>
+  </section>
+"#,
+        title = html_escape(meta.title),
+        css = CSS,
+        mode = format!("{:?}", meta.mode).to_lowercase(),
+        mode_lc = format!("{:?}", meta.mode).to_lowercase(),
+        badge_class = badge_class,
+        status_label = html_escape(&status_label),
+        observed = stats.observed,
+        denied = stats.denied,
+        denied_class = if stats.denied > 0 { "danger" } else { "" },
+        lost = stats.lost,
+        lost_class = if stats.lost > 0 { "warn" } else { "" },
+        n_exec = by_kind.exec,
+        n_open = by_kind.open,
+        n_connect = by_kind.connect,
+    );
+
+    // --- events table ---
+    html.push_str(
+        r#"  <section class="events">
+    <div class="toolbar">
+      <div class="tabs">
+        <button class="tab active" data-filter="all">All</button>
+        <button class="tab" data-filter="exec">Exec</button>
+        <button class="tab" data-filter="open">Open</button>
+        <button class="tab" data-filter="connect">Connect</button>
+        <button class="tab" data-filter="denied">Denied</button>
+      </div>
+      <input id="q" type="search" placeholder="filter by path, host, comm…" autocomplete="off">
+    </div>
+    <table class="events-table">
+      <thead><tr><th>verdict</th><th>kind</th><th>pid</th><th>comm</th><th>detail</th></tr></thead>
+      <tbody>
+"#,
+    );
+
+    for ev in &stats.samples {
+        render_event_row(&mut html, ev);
+    }
+
+    html.push_str(
+        r#"      </tbody>
+    </table>
+    <p class="hint">Showing up to 64 samples per kind. Full event stream is in the JSON log.</p>
+  </section>
+
+  <section class="policy">
+    <h2>Effective policy</h2>
+    <pre><code>"#,
+    );
+    let yaml = serde_yaml::to_string(policy).unwrap_or_else(|_| "# serialization error\n".into());
+    html.push_str(&html_escape(&yaml));
+    html.push_str(
+        r#"</code></pre>
+  </section>
+
+  <footer>
+    <div>Generated by <a href="https://github.com/bokuweb/coronarium">coronarium</a></div>
+    <div class="meta">"#,
+    );
+    let _ = write!(html, "command: <code>{}</code>", html_escape(meta.command));
+    html.push_str(
+        r#"</div>
+  </footer>
+</main>
+<script>"#,
+    );
+    html.push_str(JS);
+    html.push_str("</script>\n</body>\n</html>\n");
+
+    html
+}
+
+/// One-shot context passed in from the CLI.
+pub struct ReportMeta<'a> {
+    pub title: &'a str,
+    pub mode: crate::policy::Mode,
+    pub command: &'a str,
+}
+
+#[derive(Default)]
+struct KindCounts {
+    exec: u32,
+    open: u32,
+    connect: u32,
+}
+
+fn count_by_kind(samples: &[Event]) -> KindCounts {
+    let mut c = KindCounts::default();
+    for s in samples {
+        match s {
+            Event::Exec { .. } => c.exec += 1,
+            Event::Open { .. } => c.open += 1,
+            Event::Connect { .. } => c.connect += 1,
+        }
+    }
+    c
+}
+
+fn render_event_row(out: &mut String, ev: &Event) {
+    let denied = ev.denied();
+    let row_class = if denied { "row-denied" } else { "" };
+    let verdict_chip = if denied {
+        r#"<span class="chip chip-deny">DENY</span>"#
+    } else {
+        r#"<span class="chip chip-allow">ALLOW</span>"#
+    };
+    let (kind, kind_class, pid, comm, detail) = match ev {
+        Event::Exec {
+            pid,
+            comm,
+            filename,
+            argv0,
+            ..
+        } => (
+            "exec",
+            "chip-exec",
+            *pid,
+            comm.clone(),
+            format!(
+                r#"<code>{}</code> <span class="muted">({})</span>"#,
+                html_escape(filename),
+                html_escape(argv0)
+            ),
+        ),
+        Event::Open {
+            pid,
+            comm,
+            filename,
+            flags,
+            ..
+        } => (
+            "open",
+            "chip-open",
+            *pid,
+            comm.clone(),
+            format!(
+                r#"<code>{}</code> <span class="muted">flags=0x{:x}</span>"#,
+                html_escape(filename),
+                flags
+            ),
+        ),
+        Event::Connect {
+            pid,
+            comm,
+            daddr,
+            dport,
+            ..
+        } => (
+            "connect",
+            "chip-connect",
+            *pid,
+            comm.clone(),
+            format!(r#"<code>{}:{}</code>"#, html_escape(daddr), dport),
+        ),
+    };
+    let searchable = format!("{kind} {pid} {comm} {detail}").to_lowercase();
+    let _ = write!(
+        out,
+        r#"        <tr class="{row_class}" data-kind="{kind}" data-denied="{denied}" data-search="{search}">
+          <td>{verdict_chip}</td>
+          <td><span class="chip {kind_class}">{kind}</span></td>
+          <td class="num">{pid}</td>
+          <td><code>{comm}</code></td>
+          <td>{detail}</td>
+        </tr>
+"#,
+        row_class = row_class,
+        kind = kind,
+        denied = denied,
+        search = html_escape(&searchable),
+        verdict_chip = verdict_chip,
+        kind_class = kind_class,
+        pid = pid,
+        comm = html_escape(&comm),
+        detail = detail,
+    );
+}
+
+fn html_escape(s: &str) -> String {
+    let mut out = String::with_capacity(s.len());
+    for c in s.chars() {
+        match c {
+            '<' => out.push_str("&lt;"),
+            '>' => out.push_str("&gt;"),
+            '&' => out.push_str("&amp;"),
+            '"' => out.push_str("&quot;"),
+            '\'' => out.push_str("&#39;"),
+            _ => out.push(c),
+        }
+    }
+    out
+}
+
+const CSS: &str = r#"
+:root {
+  --bg: #f7f8fa;
+  --surface: #ffffff;
+  --surface-2: #f1f3f7;
+  --text: #18181b;
+  --text-muted: #71717a;
+  --border: #e4e4e7;
+  --accent: #6366f1;
+  --accent-soft: #e0e7ff;
+  --danger: #dc2626;
+  --danger-soft: #fee2e2;
+  --warn: #d97706;
+  --warn-soft: #fef3c7;
+  --ok: #059669;
+  --ok-soft: #d1fae5;
+  --shadow: 0 1px 2px rgba(0,0,0,.04), 0 4px 12px rgba(0,0,0,.04);
+  --mono: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
+  --sans: system-ui, -apple-system, "Segoe UI", Roboto, "Helvetica Neue", Inter, sans-serif;
+}
+@media (prefers-color-scheme: dark) {
+  :root {
+    --bg: #09090b;
+    --surface: #18181b;
+    --surface-2: #27272a;
+    --text: #fafafa;
+    --text-muted: #a1a1aa;
+    --border: #27272a;
+    --accent: #818cf8;
+    --accent-soft: #312e81;
+    --danger: #f87171;
+    --danger-soft: #450a0a;
+    --warn: #fbbf24;
+    --warn-soft: #451a03;
+    --ok: #34d399;
+    --ok-soft: #064e3b;
+    --shadow: 0 1px 2px rgba(0,0,0,.4), 0 4px 12px rgba(0,0,0,.3);
+  }
+}
+
+* { box-sizing: border-box; }
+html, body { margin: 0; padding: 0; }
+body {
+  font-family: var(--sans);
+  background: var(--bg);
+  color: var(--text);
+  font-size: 14px;
+  line-height: 1.5;
+  -webkit-font-smoothing: antialiased;
+}
+code, pre { font-family: var(--mono); font-size: 0.92em; }
+a { color: var(--accent); text-decoration: none; }
+a:hover { text-decoration: underline; }
+
+.top {
+  position: sticky; top: 0; z-index: 10;
+  display: flex; justify-content: space-between; align-items: center;
+  padding: 16px 32px;
+  background: var(--surface);
+  border-bottom: 1px solid var(--border);
+  backdrop-filter: saturate(180%) blur(12px);
+}
+.brand { display: flex; gap: 10px; align-items: center; }
+.brand .logo { font-size: 20px; }
+.brand .title { font-weight: 700; font-size: 17px; letter-spacing: -0.01em; }
+.badges { display: flex; gap: 8px; flex-wrap: wrap; }
+
+.badge {
+  display: inline-flex; align-items: center;
+  padding: 4px 10px;
+  border-radius: 999px;
+  font-size: 12px; font-weight: 500;
+  background: var(--surface-2);
+  border: 1px solid var(--border);
+  color: var(--text-muted);
+}
+.mode-block { background: var(--accent-soft); color: var(--accent); border-color: transparent; }
+.mode-audit { background: var(--surface-2); color: var(--text-muted); }
+.badge-ok { background: var(--ok-soft); color: var(--ok); border-color: transparent; }
+.badge-danger { background: var(--danger-soft); color: var(--danger); border-color: transparent; font-weight: 600; }
+.badge-warn { background: var(--warn-soft); color: var(--warn); border-color: transparent; }
+
+main { max-width: 1200px; margin: 0 auto; padding: 32px; }
+
+.stats {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 16px;
+  margin-bottom: 32px;
+}
+.card {
+  background: var(--surface);
+  border: 1px solid var(--border);
+  border-radius: 12px;
+  padding: 20px;
+  box-shadow: var(--shadow);
+}
+.card-label {
+  font-size: 11px;
+  text-transform: uppercase;
+  letter-spacing: 0.08em;
+  color: var(--text-muted);
+  font-weight: 600;
+  margin-bottom: 8px;
+}
+.card-value {
+  font-size: 34px; font-weight: 700;
+  letter-spacing: -0.02em;
+}
+.card-value.danger { color: var(--danger); }
+.card-value.warn { color: var(--warn); }
+.card-kinds { display: flex; flex-wrap: wrap; gap: 6px; margin-top: 4px; }
+
+.chip {
+  display: inline-flex; align-items: center;
+  padding: 2px 8px;
+  border-radius: 6px;
+  font-size: 11px; font-weight: 600;
+  background: var(--surface-2);
+  color: var(--text-muted);
+  font-family: var(--mono);
+  letter-spacing: 0.02em;
+}
+.chip-allow { background: var(--ok-soft); color: var(--ok); }
+.chip-deny { background: var(--danger-soft); color: var(--danger); }
+.chip-exec { background: #ddd6fe; color: #6d28d9; }
+.chip-open { background: #bfdbfe; color: #1d4ed8; }
+.chip-connect { background: #fbcfe8; color: #be185d; }
+@media (prefers-color-scheme: dark) {
+  .chip-exec { background: #4c1d95; color: #ddd6fe; }
+  .chip-open { background: #1e3a8a; color: #bfdbfe; }
+  .chip-connect { background: #831843; color: #fbcfe8; }
+}
+
+.events { margin-bottom: 32px; }
+.toolbar {
+  display: flex; justify-content: space-between; gap: 16px;
+  align-items: center; margin-bottom: 12px; flex-wrap: wrap;
+}
+.tabs { display: flex; gap: 4px; background: var(--surface-2); padding: 3px; border-radius: 8px; }
+.tab {
+  border: 0; background: transparent; cursor: pointer;
+  padding: 6px 14px; font-size: 13px; font-weight: 500;
+  border-radius: 6px; color: var(--text-muted);
+  font-family: var(--sans);
+  transition: background 120ms, color 120ms;
+}
+.tab:hover { color: var(--text); }
+.tab.active { background: var(--surface); color: var(--text); box-shadow: var(--shadow); }
+#q {
+  padding: 8px 14px;
+  min-width: 260px;
+  border: 1px solid var(--border);
+  background: var(--surface);
+  color: var(--text);
+  border-radius: 8px;
+  font-size: 13px;
+  font-family: var(--sans);
+  outline: none;
+}
+#q:focus { border-color: var(--accent); box-shadow: 0 0 0 3px var(--accent-soft); }
+
+.events-table {
+  width: 100%; border-collapse: collapse;
+  background: var(--surface);
+  border: 1px solid var(--border);
+  border-radius: 12px;
+  overflow: hidden;
+  box-shadow: var(--shadow);
+}
+.events-table th, .events-table td {
+  padding: 10px 16px; text-align: left;
+  border-bottom: 1px solid var(--border);
+  vertical-align: middle;
+}
+.events-table th {
+  background: var(--surface-2);
+  font-size: 11px; text-transform: uppercase;
+  letter-spacing: 0.08em; color: var(--text-muted);
+  font-weight: 600;
+}
+.events-table tr:last-child td { border-bottom: 0; }
+.events-table tr.row-denied { background: color-mix(in srgb, var(--danger-soft) 40%, transparent); }
+.events-table .num { font-family: var(--mono); color: var(--text-muted); }
+.events-table code { background: var(--surface-2); padding: 2px 6px; border-radius: 4px; }
+.muted { color: var(--text-muted); }
+.hint { color: var(--text-muted); font-size: 12px; margin-top: 8px; }
+
+.policy pre {
+  background: var(--surface);
+  border: 1px solid var(--border);
+  border-radius: 12px;
+  padding: 20px;
+  overflow: auto;
+  margin: 0;
+  box-shadow: var(--shadow);
+}
+.policy h2 { font-size: 16px; margin: 0 0 12px 0; letter-spacing: -0.01em; }
+
+footer {
+  margin-top: 48px;
+  padding: 24px 0;
+  border-top: 1px solid var(--border);
+  color: var(--text-muted);
+  font-size: 12px;
+  display: flex; justify-content: space-between; flex-wrap: wrap; gap: 8px;
+}
+footer code { background: var(--surface-2); padding: 2px 6px; border-radius: 4px; }
+"#;
+
+const JS: &str = r#"
+(function() {
+  const rows = document.querySelectorAll('.events-table tbody tr');
+  const tabs = document.querySelectorAll('.tab');
+  const q = document.getElementById('q');
+  let filter = 'all';
+
+  function apply() {
+    const needle = q.value.trim().toLowerCase();
+    for (const row of rows) {
+      const kind = row.dataset.kind;
+      const denied = row.dataset.denied === 'true';
+      const hay = row.dataset.search || '';
+      let kindOk = filter === 'all' || filter === kind || (filter === 'denied' && denied);
+      let qOk = !needle || hay.includes(needle);
+      row.style.display = (kindOk && qOk) ? '' : 'none';
+    }
+  }
+
+  tabs.forEach(t => t.addEventListener('click', () => {
+    tabs.forEach(x => x.classList.remove('active'));
+    t.classList.add('active');
+    filter = t.dataset.filter;
+    apply();
+  }));
+  q.addEventListener('input', apply);
+})();
+"#;
