+++ crates/coronarium-core/src/lib.rs
@@ -1,0 +1,25 @@
+//! Shared, platform-neutral building blocks for the coronarium family:
+//! - [`events`]: the audit event schema (Exec / Open / Connect)
+//! - [`stats`]: aggregate counters + per-kind sample buffer
+//! - [`policy`]: YAML/JSON policy loader
+//! - [`matcher`]: userspace-side file / exec deny matching
+//! - [`html`]: self-contained HTML report renderer
+//! - [`report`]: JSON log + `$GITHUB_STEP_SUMMARY` writer
+//!
+//! This crate exists so the Linux (`coronarium`) and Windows
+//! (`coronarium-win`) binaries can emit **identical** JSON logs and HTML
+//! reports. Anything that needs eBPF / ETW / OS-specific APIs stays in
+//! the respective binary crates.
+
+pub mod events;
+pub mod html;
+pub mod matcher;
+pub mod policy;
+pub mod report;
+pub mod stats;
+
+pub use events::Event;
+pub use policy::{
+    DefaultDecision, FilePolicy, Mode, NetRule, NetworkPolicy, Policy, ProcessPolicy,
+};
+pub use stats::Stats;
