+++ crates/coronarium-win/Cargo.toml
@@ -5,14 +5,20 @@ edition = "2024"
 license = "MIT OR Apache-2.0"
 description = "ETW-based audit counterpart to coronarium (Windows). Experimental."
 
-# Own workspace so the Windows-specific deps (ferrisetw, etc.) don't
-# pollute the main Linux workspace.
+# Own workspace so the Windows-specific deps don't pollute the main Linux
+# workspace.
 [workspace]
 
 [dependencies]
 anyhow = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+serde_yaml = "0.9"
+log = "0.4"
+env_logger = "0.11"
+clap = { version = "4", features = ["derive", "env"] }
+coronarium-core = { path = "../coronarium-core" }
+coronarium-common = { path = "../coronarium-common", features = ["user"] }
 
 [target.'cfg(windows)'.dependencies]
 ferrisetw = "1.2"
