+++ crates/coronarium/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/main.rs"
 
 [dependencies]
 coronarium-common = { path = "../coronarium-common", features = ["user"] }
+coronarium-core = { path = "../coronarium-core" }
 anyhow = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
