+++ crates/coronarium/src/main.rs
@@ -8,7 +8,6 @@ mod html;
 mod loader;
 mod matcher;
 mod policy;
-mod report;
 mod resolve;
 
 use anyhow::Result;
